pyproject.toml
[project]
name = "website-rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=1.0.7",
    "crawl4ai==0.7.7",
    "pydantic-ai>=0.1.8",
    "python-dotenv==1.1.0",
    "streamlit>=1.45.0",
    "playwright",
    "more-itertools>=10.8.0",
    "sentence-transformers>=5.1.2",
]
